requirements.txt
python-dotenv==1.0.0
requests==2.28.2
peewee==3.16.0
pyTelegramBotAPI==4.10.0
python-telegram-bot-calendar==1.0.5
